Add object augmentation and .obj support

+++ headers/Segmentation.hpp
@@ -18,7 +18,7 @@ namespace arfs
         static std::vector<cv::Point> recognizeTag(const cv::Mat& frame, std::vector<std::vector<cv::Point>> candidates, std::array<int, 64> code);
         static std::array<int, 64> getARTagCode(const cv::Mat& tag_img);
         static cv::Mat getProjectionMatrix(const cv::Mat& homography, const cv::Matx33d& intrinsicMatrix);
-        static std::vector<cv::Point2d> projectPoint(const std::vector<cv::Point3d>& points, const cv::Mat& projectionMatrix);
+        static std::vector<cv::Point2i> projectPoint(const std::vector<cv::Point3d>& points, const cv::Mat& projectionMatrix);
         static void augmentObject(const OBJLoader& obj, const cv::Mat& frame, const std::vector<cv::Point>& tag);
     private:
         static cv::Mat threshold(const cv::Mat& img);
